+++ src/input.rs
@@ -13,7 +13,7 @@ const BUFFER_SIZE: usize = 1024;
 
 const ESCAPE: u8 = 0o33;
 
-pub fn get_input() -> impl Iterator<Item = io::Result<TuiEvent>> {
+pub fn remap_dev_tty_to_stdin() {
     // The readline library we use, rustyline, always gets its input from STDIN.
     // If jless accepts its input from STDIN, then rustyline can't accept input.
     // To fix this, we open up /dev/tty, and remap it to STDIN, as suggested in
@@ -30,8 +30,14 @@ pub fn get_input() -> impl Iterator<Item = io::Result<TuiEvent>> {
         let path = std::ffi::CString::new("r").unwrap();
         let _ = libc::freopen(filename.as_ptr(), path.as_ptr(), libc_stdhandle::stdin());
     }
+}
 
+pub fn get_input() -> impl Iterator<Item = io::Result<TuiEvent>> {
     let (sigwinch_read, sigwinch_write) = UnixStream::pair().unwrap();
+    // NOTE: This overrides the SIGWINCH handler registered by rustyline.
+    // We should maybe get a reference to the existing signal handler
+    // and call it when appropriate, but it seems to only be used to handle
+    // line wrapping, and it seems to work fine without it.
     pipe::register(SIGWINCH, sigwinch_write).unwrap();
     TuiInput::new(stdin(), sigwinch_read)
 }